+++ change_analyzer/sequences_diff.py
@@ -2,258 +2,165 @@ import argparse
 import glob
 import json
 import os
+from os.path import basename, dirname
 import logging
-from typing import Tuple, List
+import re
+from typing import Tuple, List, Dict, Union
 
+import PIL.Image
 import numpy as np
 import pandas as pd
-import matplotlib.image as mpimg
+from PIL import Image
+from PIL import ImageDraw
 
+from xmldiff import main as diffmain
+import xml.etree.ElementTree as ET
 from datetime import datetime
 from jinja2 import Environment, FileSystemLoader
 from change_analyzer.wrappers.sequence_recorder import SequenceRecorder
 
 
 class SequencesDiff:
-
-    def __init__(self, sequence1_folder: str = None, sequence2_folder: str = None) -> None:
-        # - If sequence1 and sequence2 folders are not provided, we test the last two sequences
-        # - If only sequence1 folder is provided, we test it against the last valid sequence
+    def __init__(self, sequence1_file: str, sequence2_file: str) -> None:
         self._logger = logging.getLogger(__name__)
-
-        self.sequence1_file = None
-        self.sequence2_file = None
-        self.sequence1_date = None
-        self.sequence2_date = None
-        self.report_file = None
-        self.expected_sequence_id = None
-        self.actual_sequence_id = None
-        self.expected_images = None
-        self.actual_images = None
-        self.verdicts = []
-        self.comments = []
-        self.expected_steps = []
-        self.actual_steps = []
-        self.report_date = None
+        self.report_date = datetime.now().strftime("%Y_%m_%d-%H_%M_%S")
+        self.sequence1_date = basename(dirname(sequence1_file))
+        self.sequence2_date = basename(dirname(sequence2_file))
+        self.expected_sequence_id = basename(sequence1_file).split('.')[0]
+        self.actual_sequence_id = basename(sequence2_file).split('.')[0]
+        self.report_folder = os.path.join(dirname(sequence1_file), "comparisons", self.report_date)
+        os.makedirs(self.report_folder, exist_ok=True)
         self.df_merged = pd.DataFrame()
 
-        self.sequence1_folder = sequence1_folder
-        self.sequence2_folder = sequence2_folder
-        self.update_sequence_files()
-        self.update_report_file_path()
-        self.update_sequences_date()
-        self.create_comparisons_folder()
-        self.step_count = 0
+        self._update_merged_df(sequence1_file, sequence2_file)
+        self._write_to_report(os.path.join(self.report_folder, 'index.html'))
 
-        self.validate_replay()
-
-        self.write_to_report()
-
-    def find_csv_file_within_folder(self, folder: str) -> str:
-
-        try:
-            csv_file = glob.glob(os.path.join(os.getcwd(), "recordings", folder, "*.csv"))[0]
-            self._logger.info("csv file found ", csv_file)
-            return csv_file
-        except Exception as e:
-            self._logger.info("We couldn't find any csv file within the folder due to an exception")
-            self._logger.info(e)
-
-    def find_last_two_valid_folders(self) -> Tuple[str, str]:
-        """Find the last two folders which have a csv file"""
-        folders_found = 0
-        csv_file_paths = []
-        recordings_folder = os.path.join(os.path.abspath(os.getcwd()), 'recordings')
-        os.chdir(recordings_folder)
-        folders = list(filter(os.path.isdir, os.listdir(recordings_folder)))
-        folders = [os.path.join(recordings_folder, f) for f in folders]  # add path to each file
-        folders.sort(key=lambda x: os.path.getmtime(x))
-        folders.reverse()
-
-        for folder in folders:
-            csv_file_path = self.find_csv_file_within_folder(folder)
-            if csv_file_path:
-                folders_found += 1
-                csv_file_paths.append(csv_file_path)
-            if folders_found == 2:
-                break
-
-        return csv_file_paths[0], csv_file_paths[1]
-
-    def update_sequences_date(self):
-        sequence1_folder = os.path.dirname(self.sequence1_file)
-        sequence2_folder = os.path.dirname(self.sequence2_file)
-        self.sequence1_date = os.path.basename(sequence1_folder)
-        self.sequence2_date = os.path.basename(sequence2_folder)
-
-    def update_sequence_files(self):
-
-        if not self.sequence1_folder and not self.sequence2_folder:
-            self.sequence1_file, self.sequence2_file = self.find_last_two_valid_folders()
-            self._logger.info("We have no sequence defined.")
-
-        if self.sequence1_folder and not self.sequence2_folder:
-            self.sequence1_file = self.find_csv_file_within_folder(self.sequence1_folder)
-            self.sequence2_file, _ = self.find_last_two_valid_folders()
-            self._logger.info("We have only expected sequence defined.")
-
-        if self.sequence1_folder and self.sequence2_folder:
-            self.sequence1_file = self.find_csv_file_within_folder(self.sequence1_folder)
-            self.sequence2_file = self.find_csv_file_within_folder(self.sequence2_folder)
-            self._logger.info("We have both sequences defined.")
-
-        self._logger.info("We use sequence1: ", self.sequence1_file)
-        self._logger.info("We use sequence2: ", self.sequence2_file)
-
-    def create_comparisons_folder(self):
-        folder = os.path.join(os.path.dirname(self.sequence1_file), "comparisons")
-        os.makedirs(folder, exist_ok=True)
-
-    def update_report_file_path(self):
-        """ The report file will always be within the first sequence folder"""
-        self.expected_sequence_id = os.path.basename(self.sequence1_file).split('.')[0]
-        self.actual_sequence_id = os.path.basename(self.sequence2_file).split('.')[0]
-        self.report_date = datetime.now().strftime("%Y_%m_%d-%H_%M_%S")
-        file_name = f'{self.expected_sequence_id}_vs_{self.actual_sequence_id}_{self.report_date}.html'
-        self.report_file = os.path.join(os.path.dirname(self.sequence1_file), "comparisons", file_name)
-
-    def validate_replay(self):
-        df_sequence1 = pd.read_csv(self.sequence1_file)
-        df_sequence2 = pd.read_csv(self.sequence2_file)
+    def _update_merged_df(self, sequence1_file, sequence2_file) -> None:
+        """Update merged dataframe using expected and actual dataframes"""
         self.df_merged = pd.merge(
-            df_sequence1,
-            df_sequence2,
+            pd.read_csv(sequence1_file),
+            pd.read_csv(sequence2_file),
             left_index=True,
             right_index=True,
             suffixes=('_expected', '_actual')
         )
 
-        self.expected_steps = list(self.df_merged[f'{SequenceRecorder.COL_ACTION_TO_PERFORM}_expected'])
-        self.actual_steps = list(self.df_merged[f'{SequenceRecorder.COL_ACTION_TO_PERFORM}_actual'])
-        valid = self.validate_steps(self.expected_steps, self.actual_steps)
+        # Add the first step into DF, based on current first row of DF
+        first_step_data = {
+            f'{SequenceRecorder.COL_SEQUENCE_ID}_expected': self.df_merged[f'{SequenceRecorder.COL_SEQUENCE_ID}_expected'][0],
+            f'{SequenceRecorder.COL_SEQUENCE_ID}_actual': self.df_merged[f'{SequenceRecorder.COL_SEQUENCE_ID}_actual'][0],
+            f'{SequenceRecorder.COL_PAGE_SOURCE_AFTER}_actual': self.df_merged[f'{SequenceRecorder.COL_PAGE_SOURCE_BEFORE}_actual'][0],
+            f'{SequenceRecorder.COL_PAGE_SOURCE_AFTER}_expected': self.df_merged[f'{SequenceRecorder.COL_PAGE_SOURCE_BEFORE}_expected'][0],
+            f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_actual': self.df_merged[f'{SequenceRecorder.COL_ACTION_IMAGE_BEFORE}_actual'][0],
+            f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_expected': self.df_merged[f'{SequenceRecorder.COL_ACTION_IMAGE_BEFORE}_expected'][0],
+            f'{SequenceRecorder.COL_ACTION_TO_PERFORM}_expected': 'open the Application',
+            f'{SequenceRecorder.COL_ACTION_TO_PERFORM}_actual': 'open the Application',
+        }
+        first_step_df = pd.DataFrame([first_step_data])
+        self.df_merged = pd.concat([first_step_df, self.df_merged], ignore_index=True)
+        self.df_merged.to_csv(os.path.join(self.report_folder, 'df.csv'), index=False)
 
-        if valid:
-            self._logger.info("Actual steps are the same as expected steps")
-            self.df_merged.apply(self.validate_steps_output, axis=1)
-        else:
-            self._logger.info("Actual steps are not the same as expected steps")
+        # Images
+        img_col = [col for col in self.df_merged.columns if "image" in col.lower()]
+        for col in img_col:
+            self.df_merged[col] = self.df_merged[col].apply(self.json_to_image)
+        self.df_merged["ImageVerdict"] = self.df_merged.apply(lambda row: "pass" if np.array_equal(
+            row[f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_expected'],
+            row[f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_actual']
+        ) else "fail", axis=1)
 
-    def validate_steps(self, expected_steps: List[str], actual_steps: List[str]) -> bool:
-        """Validate if the performed steps are the same"""
+        # XMLs
+        xml_cols = [col for col in self.df_merged.columns if "pagesource" in col.lower()]
+        for col in xml_cols:
+            self.df_merged[col] = self.df_merged[col].apply(self.encode_xml)
 
-        if len(expected_steps) != len(actual_steps):
-            self._logger.info("The amount of actual steps is not the same as the amount of performed steps")
-            self._logger.info(f"Actual steps: {len(actual_steps)}\n", actual_steps)
-            self._logger.info(f"Expected steps: {len(expected_steps)}\n", expected_steps)
-            return False
+        self.df_merged["PageSource_diff"] = self.df_merged.apply(lambda row: self._extract_diffs_as_dict(
+            row[f'{SequenceRecorder.COL_PAGE_SOURCE_AFTER}_expected'],
+            row[f'{SequenceRecorder.COL_PAGE_SOURCE_AFTER}_actual']
+        ), axis=1)
+        self.df_merged["DiffInfo"] = self.df_merged["PageSource_diff"].apply(self._get_diff_info)
 
-        if expected_steps == actual_steps:
-            self._logger.info("Actual steps are the same as expected steps\n", actual_steps)
-            return True
+        # Image highlighting
+        self.df_merged.apply(lambda row: self._draw_boundaries(
+            row["DiffInfo"],
+            row[f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_actual'],
+            row[f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_expected'],
+        ), axis=1)
+        self.df_merged["actual_screenshot_filename"] = self.df_merged.apply(lambda row: f"actual_screenshot_step{row.name + 1}.png", axis=1)
+        self.df_merged["expected_screenshot_filename"] = self.df_merged.apply(lambda row: f"expected_screenshot_step{row.name + 1}.png", axis=1)
 
-        self._logger.info("Actual steps are not the same as expected steps")
-        for i, step in enumerate(expected_steps):
-            expected_step = expected_steps[i]
-            actual_step = actual_steps[i]
-            if expected_step != actual_step:
-                self._logger.info(f"Step {i+1} is not as expected. Expected {expected_step} and found {actual_step}")
-                return False
-
-        return True
+        self.df_merged.apply(lambda row: self._save_step_images(
+            row[f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_actual'],
+            row[f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_expected'],
+            row["actual_screenshot_filename"],
+            row["expected_screenshot_filename"],
+        ), axis=1)
 
     @staticmethod
-    def validate_actual_vs_expected_images(image_actual: str, image_expected: str) -> bool:
-        """Validate actual image vs expected image of the same step"""
-        # Convert images to arrays
-        image_expected_as_array = np.array(json.loads(image_expected), dtype=np.uint8)
-        image_actual_as_array = np.array(json.loads(image_actual), dtype=np.uint8)
+    def json_to_image(img: str) -> Union[PIL.Image.Image, float]:
+        try:
+            return Image.fromarray(np.array(json.loads(img), dtype=np.uint8))
+        except:
+            return np.nan
 
-        return np.array_equal(image_expected_as_array, image_actual_as_array)
+    @staticmethod
+    def encode_xml(xml: str) -> Union[bytes, float]:
+        try:
+            return xml.encode(re.findall(r'encoding=\"(.+?)\"', xml)[0])
+        except:
+            return np.nan
 
-    def save_step_images_from_arrays(self, image_actual: str, image_expected: str, filename_actual, filename_expected):
+    def _save_step_images(self, image_actual: PIL.Image.Image, image_expected: PIL.Image.Image, filename_actual: str, filename_expected: str) -> None:
+        """Save step expected and actual images of the SUT"""
+        expected_image_filepath = os.path.join(self.report_folder, filename_expected)
+        actual_image_filepath = os.path.join(self.report_folder, filename_actual)
 
-        folder = os.path.dirname(self.report_file)
+        image_expected.save(expected_image_filepath, format="PNG")
+        image_actual.save(actual_image_filepath, format="PNG")
 
-        # Convert images to arrays
-        image_expected_as_array = np.array(json.loads(image_expected), dtype=np.uint8)
-        image_actual_as_array = np.array(json.loads(image_actual), dtype=np.uint8)
+    @staticmethod
+    def _draw_boundaries(info_list: List[Dict], image_actual_pil: PIL.Image.Image, image_expected_pil: PIL.Image.Image) -> None:
+        if len(info_list) == 0:
+            return
 
-        expected_image_filepath = os.path.join(folder, filename_expected)
-        actual_image_filepath = os.path.join(folder, filename_actual)
+        diff_type = {
+            'UpdateAttrib': 'orange',
+            'DeleteAttrib': 'red',
+            'InsertAttrib': 'green'
+        }
+        for info in info_list:
+            actual_element_boundaries = info['actual_element_boundaries']
+            expected_element_boundaries = info['expected_element_boundaries']
 
-        mpimg.imsave(expected_image_filepath, image_expected_as_array)
-        mpimg.imsave(actual_image_filepath, image_actual_as_array)
+            # Get outline color, based on diff type
+            outline_color = diff_type[info['actual_element']['type']]
 
-    def update_verdicts_and_comments(self, valid: bool):
-        if valid:
-            self.verdicts.append("pass")
-            self.comments.append("The actual screenshot is the same as the expected screenshot")
-        else:
-            self.verdicts.append("fail")
-            self.comments.append("The actual screenshot is not the same as the expected screenshot")
+            # Draw boundaries
+            draw_image_actual = ImageDraw.Draw(image_actual_pil)
+            draw_image_expected = ImageDraw.Draw(image_expected_pil)
+            draw_image_actual.polygon(actual_element_boundaries, fill=None, outline=outline_color)
+            draw_image_expected.polygon(expected_element_boundaries, fill=None, outline=outline_color)
 
-    def validate_steps_output(self, series: pd.Series):
-        """Validate if each step's output is as expected"""
+    def _write_to_report(self, report_file) -> None:
+        """Write the report file using Jinja template"""
+        template_folder = os.path.abspath(os.path.join(__file__, "../templates"))
+        template = Environment(loader=FileSystemLoader(template_folder), autoescape=True).get_template("Log_template.html")
 
-        image_expected = series[f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_expected']
-        image_actual = series[f'{SequenceRecorder.COL_ACTION_IMAGE_AFTER}_actual']
+        self.expected_images = self.df_merged["expected_screenshot_filename"]
+        self.actual_images = self.df_merged["actual_screenshot_filename"]
+        verdicts = self.df_merged["ImageVerdict"].tolist()
+        comments = ["The actual screenshot is the same as the expected screenshot"
+                    if verdict == "pass"
+                    else "The actual screenshot is not the same as the expected screenshot"
+                    for verdict in verdicts]
+        expected_steps = self.df_merged[f'{SequenceRecorder.COL_ACTION_TO_PERFORM}_expected'].tolist()
+        actual_steps = self.df_merged[f'{SequenceRecorder.COL_ACTION_TO_PERFORM}_actual'].tolist()
 
-        # Validate actual vs expected images are the same
-        valid = self.validate_actual_vs_expected_images(image_actual, image_expected)
-
-        # Save images
-        expected_image_filename = f"expected_screenshot_step{len(self.verdicts) + 2}_{self.report_date}.png"
-        actual_image_filename = f"actual_screenshot_step{len(self.verdicts) + 2}_{self.report_date}.png"
-        self.save_step_images_from_arrays(image_actual, image_expected, actual_image_filename, expected_image_filename)
-
-        # Update verdicts and comments based on the step validation
-        self.update_verdicts_and_comments(valid)
-
-    def add_initial_step_to_sequences_data_components(self):
-        """Add an initial step, Open Application, validate actual and expected images and save them"""
-
-        # Add initial data regarding the first step
-        self.expected_steps.insert(0, "open the Application")
-        self.actual_steps.insert(0, "open the Application")
-        self.expected_images.insert(0, f"expected_screenshot_step1_{self.report_date}")
-        self.actual_images.insert(0, f"actual_screenshot_step1_{self.report_date}")
-
-        # Get the images from merged DF
-        expected_image = self.df_merged[f'{SequenceRecorder.COL_ACTION_IMAGE_BEFORE}_expected'][0]
-        actual_image = self.df_merged[f'{SequenceRecorder.COL_ACTION_IMAGE_BEFORE}_actual'][0]
-
-        # Save the step images
-        expected_image_filename = f"expected_screenshot_step1_{self.report_date}.png"
-        actual_image_filename = f"actual_screenshot_step1_{self.report_date}.png"
-        self.save_step_images_from_arrays(actual_image, expected_image, actual_image_filename, expected_image_filename)
-
-        # Validate step images
-        valid = self.validate_actual_vs_expected_images(actual_image, expected_image)
-
-        if valid:
-            self.verdicts.insert(0, "pass")
-            self.comments.insert(0, "The actual screenshot is the same as the expected screenshot")
-        else:
-            self.verdicts.insert(0, "fail")
-            self.comments.insert(0, "The actual screenshot is not the same as the expected screenshot")
-
-    def write_to_report(self):
-        template_folder = os.path.abspath(os.path.join(__file__, "../../templates"))
-        file_loader = FileSystemLoader(template_folder)
-        env = Environment(loader=file_loader, autoescape=True)
-        template = env.get_template("Log_template.html")
-
-        self.expected_images = [f'expected_screenshot_step{i+2}_{self.report_date}' for i in range(len(self.expected_steps))]
-        self.actual_images = [f'actual_screenshot_step{i+2}_{self.report_date}' for i in range(len(self.actual_steps))]
-
-        # Add the very first step
-        self.add_initial_step_to_sequences_data_components()
-
-        sequences_data = zip(self.expected_steps, self.actual_steps,
+        sequences_data = zip(expected_steps, actual_steps,
                              self.expected_images, self.actual_images,
-                             self.verdicts, self.comments)
+                             verdicts, comments)
 
-        steps = [f"step{i+1}" for i in range(len(self.expected_steps))]
+        steps = [f"step{i+1}" for i in range(len(expected_steps))]
         aria_controls = ",".join(steps)
 
         # Render HTML Template String
@@ -266,9 +173,139 @@ class SequencesDiff:
                                                sequence2_date=self.sequence2_date)
 
         # Save the new HTML file
-        with open(self.report_file, "w") as f:
+        with open(report_file, "w") as f:
             f.write(html_template_string)
 
+    @staticmethod
+    def _get_attribute_value_based_on_node(page_root: ET.Element, node: str, attribute: str) -> str:
+        """Get attribute value for specific node from given page source"""
+        formatted_node = "./" + "/".join(node.split("/")[2::])
+        elem = page_root.findall(formatted_node)[0]
+        return elem.attrib[attribute]
+
+    def _extract_diffs_as_dict(self, expected_xml: str, actual_xml: str) -> Dict:
+        """Extract diffs as dictionary, using expected and actual approach"""
+        diffs_as_dict = {
+            'expected': {},
+            'actual': {}
+        }
+        # expected_xml and actual_xml can sometimes be nan, giving a TypeError
+        try:
+            expected_root = ET.ElementTree(ET.fromstring(expected_xml)).getroot()
+            actual_root = ET.ElementTree(ET.fromstring(actual_xml)).getroot()
+        except TypeError:
+            return diffs_as_dict
+
+        for diff in diffmain.diff_texts(expected_xml, actual_xml):
+            if 'ProcessId' not in diff and 'RuntimeId' not in diff:
+                diff_type = str(diff).split("(")[0]
+                expected_value = self._get_attribute_value_based_on_node(expected_root, diff.node, diff.name)
+                if diff.node not in diffs_as_dict['actual'].keys():
+                    diffs_as_dict['actual'][diff.node] = {diff.name: diff.value}
+                    diffs_as_dict['actual'][diff.node] = {'type': diff_type}
+                    diffs_as_dict['expected'][diff.node] = {diff.name: expected_value}
+                else:
+                    diffs_as_dict['actual'][diff.node][diff.name] = diff.value
+                    diffs_as_dict['expected'][diff.node][diff.name] = expected_value
+
+        # Update each node from dict with expected and actual x,y,w,h if they are not already present
+        for k, v in diffs_as_dict['actual'].items():
+            expected_element = diffs_as_dict['expected'][k]
+            actual_element = diffs_as_dict['actual'][k]
+            if 'x' not in diffs_as_dict['actual'][k].keys():
+                expected_x = self._get_attribute_value_based_on_node(expected_root, k, 'x')
+                actual_x = self._get_attribute_value_based_on_node(actual_root, k, 'x')
+                expected_element['x'] = expected_x
+                actual_element['x'] = actual_x
+            if 'y' not in diffs_as_dict['actual'][k].keys():
+                expected_y = self._get_attribute_value_based_on_node(expected_root, k, 'y')
+                actual_y = self._get_attribute_value_based_on_node(actual_root, k, 'y')
+                expected_element['y'] = expected_y
+                actual_element['y'] = actual_y
+            if 'width' not in diffs_as_dict['actual'][k].keys():
+                expected_width = self._get_attribute_value_based_on_node(expected_root, k, 'width')
+                actual_width = self._get_attribute_value_based_on_node(actual_root, k, 'width')
+                expected_element['width'] = expected_width
+                actual_element['width'] = actual_width
+            if 'height' not in diffs_as_dict['actual'][k].keys():
+                expected_height = self._get_attribute_value_based_on_node(expected_root, k, 'height')
+                actual_height = self._get_attribute_value_based_on_node(actual_root, k, 'height')
+                expected_element['height'] = expected_height
+                actual_element['height'] = actual_height
+
+        return diffs_as_dict
+
+    def _get_diff_info(self, diff_results: Dict) -> List[Dict]:
+        if diff_results['expected'] == {} or diff_results['actual'] == {}:
+            return []
+
+        step_diff_info = []
+        for element in diff_results['actual'].keys():
+            expected_element = diff_results['expected'][element]
+            actual_element = diff_results['actual'][element]
+
+            expected_x = expected_element['x']
+            actual_x = actual_element['x']
+            expected_y = expected_element['y']
+            actual_y = actual_element['y']
+            expected_width = expected_element['width']
+            actual_width = actual_element['width']
+            expected_height = expected_element['height']
+            actual_height = actual_element['height']
+
+            expected_box_coordinates = self._get_box_coordinates(expected_x, expected_y,
+                                                                 expected_width, expected_height)
+            actual_box_coordinates = self._get_box_coordinates(actual_x, actual_y,
+                                                               actual_width, actual_height)
+            info = f'Actual ({actual_box_coordinates}) vs expected ({expected_box_coordinates})'
+
+            step_diff_info.append({
+                'actual_element': actual_element,
+                'expected_element': expected_element,
+                'actual_element_boundaries': actual_box_coordinates,
+                'expected_element_boundaries': expected_box_coordinates,
+                'info': info
+            })
+
+        return step_diff_info
+
+    @staticmethod
+    def _get_box_coordinates(x: str, y: str, w: str, h: str) -> List[Tuple[int, int]]:
+        """Get boundary coordinates based on x, y, width and height"""
+        x = int(x)
+        y = int(y)
+        w = int(w)
+        h = int(h)
+        return [(x, y), (x+w, y), (x+w, y+h), (x, y+h)]
+
+
+def find_csv_file_within_folder(folder: str) -> str:
+    """"Find the sequence csv file within the associated folder, assuming there is only one csv file present"""
+    try:
+        return glob.glob(os.path.join(os.getcwd(), "recordings", folder, "*.csv"))[0]
+    except:
+        pass
+
+def find_last_two_valid_folders() -> Tuple[str, str]:
+    """Find the last two folders which have a csv file"""
+    folders_found = 0
+    csv_file_paths = []
+    recordings_folder = os.path.join(os.path.abspath(os.getcwd()), 'recordings')
+    os.chdir(recordings_folder)
+    folders = list(filter(os.path.isdir, os.listdir(recordings_folder)))
+    folders = [os.path.join(recordings_folder, f) for f in folders]  # add path to each file
+    folders.sort(key=lambda x: os.path.getmtime(x))
+    folders.reverse()
+
+    for folder in folders:
+        csv_file_path = find_csv_file_within_folder(folder)
+        if csv_file_path:
+            folders_found += 1
+            csv_file_paths.append(csv_file_path)
+        if folders_found == 2:
+            break
+
+    return csv_file_paths[0], csv_file_paths[1]
 
 def main():
     parser = argparse.ArgumentParser()
@@ -284,7 +321,20 @@ def main():
     )
     args = parser.parse_args()
 
-    SequencesDiff(sequence1_folder=args.sequence1_folder, sequence2_folder=args.sequence2_folder)
+    sequence1_file = ""
+    sequence2_file = ""
+    if not args.sequence1_folder and not args.sequence2_folder:
+        sequence1_file, sequence2_file = find_last_two_valid_folders()
+
+    if args.sequence1_folder and not args.sequence2_folder:
+        sequence1_file = find_csv_file_within_folder(args.sequence1_folder)
+        sequence2_file, _ = find_last_two_valid_folders()
+
+    if args.sequence1_folder and args.sequence2_folder:
+        sequence1_file = find_csv_file_within_folder(args.sequence1_folder)
+        sequence2_file = find_csv_file_within_folder(args.sequence2_folder)
+
+    SequencesDiff(sequence1_file, sequence2_file)
 
 
 if __name__ == "__main__":