+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 <!--next-version-placeholder-->
 
+## v0.2.4 (2021-10-14)
+### Fix
+* Fix the case when xmls were nan ([`df6d18e`](https://github.com/F-Secure/change-analyzer/commit/df6d18e84410384f8cdda088df67e0c2c6b2bedb))
+
+## v0.2.3 (2021-10-14)
+### Fix
+* Fix setup.py readme import issue ([`5d5597c`](https://github.com/F-Secure/change-analyzer/commit/5d5597c099d8729df7d84a779acfcce9eb2b5941))
+
+## v0.2.2 (2021-10-13)
+### Fix
+* Fix large width for report images on web platform ([`a33aafa`](https://github.com/F-Secure/change-analyzer/commit/a33aafa09fa4bd027fbb663afb511c240e777172))
+
+### Documentation
+* Fix of acknowledgement section to README.md ([`6379688`](https://github.com/F-Secure/change-analyzer/commit/6379688ba3eec11239002559e7efd1db4ece7f16))
+* Added acknowledgement section to README.md ([`1f9f0b7`](https://github.com/F-Secure/change-analyzer/commit/1f9f0b7959fe47f6b84c3d0b08e1f68833b1c332))
+
+## v0.2.1 (2021-10-13)
+### Fix
+* Fix dataframe merge ([`1c41875`](https://github.com/F-Secure/change-analyzer/commit/1c41875e2fbf12b0ed78590c4e40a8732000e6cd))
+
+## v0.2.0 (2021-10-12)
+### Feature
+* Add highlighting diff on elements based on xmldiff ([`00c0c53`](https://github.com/F-Secure/change-analyzer/commit/00c0c53a82662546f317e44c863bf4a8d58d80c9))
+
+## v0.1.1 (2021-10-12)
+### Fix
+* Add template to install package ([`990bfd1`](https://github.com/F-Secure/change-analyzer/commit/990bfd1027a9fd36239b3495ef4408b2b99333f6))
+
+### Documentation
+* Add support for Conda environment ([`b3d3ee9`](https://github.com/F-Secure/change-analyzer/commit/b3d3ee952f13e20f6268f61e975c6c3c3318b065))
+
 ## v0.1.0 (2021-10-08)
 ### Feature
 * Initialize project to github ([`e352602`](https://github.com/F-Secure/change-analyzer/commit/e3526023eceb7de77ef28b64afc13ca65b5feccb))